bug curd

+++ src/components/Admin/Cards/CardCategory.js
@@ -11,35 +11,48 @@ export default function CardCategory({ color }) {
   const alert = useAlert();
   const [loading, setLoading] = useState(false);
   const [data, setData] = useState([]);
+  const [empty, setEmpty] = useState(false);
 
   useEffect(() => {
     setLoading(true);
-    axios.get("https://api.sarafdesign.com/category").then((res) => {
-      setData(res.data);
-      setLoading(false);
-    });
-  }, []);
+    axios
+      .get("https://api.sarafdesign.com/category")
+      .then((res) => {
+        setData(res.data);
+      })
+      .catch((error) => {
+        setData([]);
+        setEmpty(true);
+      });
+    setLoading(false);
+  }, [data]);
+
+  console.log("ini hasil data " + data);
 
   let handleDelete = (e) => {
-    setLoading(true);
-    axios.delete(`https://api.sarafdesign.com/category/${e}`).then((res) => {
-      setTimeout(() => {
-        alert.show("Category Successfully Deleted");
-        window.location.reload();
-      }, 5000);
+    if (window.confirm("Apakah anda yakin ingin menghapus?")) {
+      setLoading(true);
+      axios.delete(`https://api.sarafdesign.com/category/${e}`).then((res) => {
+        console.log("inihasil delete " + res);
 
-      // axios
-      //   .get(`https://api.sarafdesign.com/images/gallery/${id}`)
-      //   .then((res2) => {
-      //     setData(res2.data);
-      //     setTimeout(() => {
-      //       alert("Kehapus");
-      //       window.location.reload();
-      //     }, 5000);
-      //   });
-    });
-    setLoading(false);
-    // console.log(e, x);
+        alert.show("Category Successfully Deleted");
+        // window.location.reload();
+
+        // axios
+        //   .get(`https://api.sarafdesign.com/images/gallery/${id}`)
+        //   .then((res2) => {
+        //     setData(res2.data);
+        //     setTimeout(() => {
+        //       alert("Kehapus");
+        //       window.location.reload();
+        //     }, 5000);
+        //   });
+      }, []);
+      setLoading(false);
+      // console.log(e, x);
+    } else {
+    }
+    // e.preventDefault();
   };
 
   return (
@@ -111,48 +124,56 @@ export default function CardCategory({ color }) {
               </tr>
             </thead>
             <tbody>
-              {loading && (
-                <div>
-                  <Loader
-                    className="flex items-center justify-center mx-auto text-center mt-10 mb-10"
-                    type="Oval"
-                    color="#00BFFF"
-                    height={80}
-                    width={80}
-                  />
+              {empty ? (
+                <div className="w-full justify-center items-center flex flex-col p-5">
+                  data kosong
                 </div>
-              )}
-              {!loading &&
-                data.map((x) => (
-                  <tr>
-                    <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 text-left font-bold">
-                      1
-                    </td>
+              ) : (
+                <>
+                  {loading && (
+                    <div>
+                      <Loader
+                        className="flex items-center justify-center mx-auto text-center mt-10 mb-10"
+                        type="Oval"
+                        color="#00BFFF"
+                        height={80}
+                        width={80}
+                      />
+                    </div>
+                  )}
+                  {!loading &&
+                    data.map((x) => (
+                      <tr>
+                        <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 text-left font-bold">
+                          1
+                        </td>
 
-                    <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 text-left font-bold">
-                      {x.category_nama}
-                    </td>
-                    <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs  p-4">
-                      <div className="flex">
-                        <Link to={`/admin/editcategory/${x.id_category}`}>
-                          <button
-                            className="bg-yellow-500 text-white active:bg-blue-600 font-bold  text-xs px-4 py-2 rounded shadow hover:shadow-md outline-none focus:outline-none mr-1 ease-linear transition-all duration-150"
-                            type="button"
-                          >
-                            Edit
-                          </button>
-                        </Link>
-                        <button
-                          className="bg-red-500 text-white active:bg-blue-600 font-bold  text-xs px-4 py-2 rounded shadow hover:shadow-md outline-none focus:outline-none mr-1 ease-linear transition-all duration-150"
-                          type="button"
-                          onClick={() => handleDelete(x.id_category)}
-                        >
-                          Delete
-                        </button>
-                      </div>
-                    </td>
-                  </tr>
-                ))}
+                        <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 text-left font-bold">
+                          {x.category_nama}
+                        </td>
+                        <td className="border-t-0 px-6 align-middle border-l-0 border-r-0 text-xs  p-4">
+                          <div className="flex">
+                            <Link to={`/admin/editcategory/${x.id_category}`}>
+                              <button
+                                className="bg-yellow-500 text-white active:bg-blue-600 font-bold  text-xs px-4 py-2 rounded shadow hover:shadow-md outline-none focus:outline-none mr-1 ease-linear transition-all duration-150"
+                                type="button"
+                              >
+                                Edit
+                              </button>
+                            </Link>
+                            <button
+                              className="bg-red-500 text-white active:bg-blue-600 font-bold  text-xs px-4 py-2 rounded shadow hover:shadow-md outline-none focus:outline-none mr-1 ease-linear transition-all duration-150"
+                              type="button"
+                              onClick={() => handleDelete(x.id_category)}
+                            >
+                              Delete
+                            </button>
+                          </div>
+                        </td>
+                      </tr>
+                    ))}
+                </>
+              )}
             </tbody>
           </table>
         </div>
